feat: changement structure des données et du module
* enlève les imports circulaires de assemble.js vers les fichiers data
* réduit de 70% la taille de data.json
