+++ .gitignore
@@ -102,3 +102,9 @@ dist
 
 # TernJS port file
 .tern-port
+
+# Custom build output
+data/cpf.json
+data/cepages.json
+dist/
+*.d.ts
